requirements.txt
Django==3.2.13
APScheduler==3.9.1
PyMySQL==1.0.2
django-apscheduler==0.6.2
amadeus==5.3.1
python-dotenv==0.20.0
django-dotenv==1.4.2
requests==2.27.1
uWSGI==2.0.20